+++ lib/index.js
@@ -1,0 +1,2 @@
+export { default as ReactCrop } from './ReactCrop';
+export { default as withCrop } from './withCrop';
