+++ webpack.config.js
@@ -1,7 +1,7 @@
 const path = require('path');
 
 module.exports = {
-  entry: './lib/ReactCrop',
+  entry: './lib/index',
   output: {
     path: path.join(__dirname, 'dist'),
     library: 'ReactCrop',
